Migrate lists test to wit-bindgen test
