+++ tests/runtime/main.rs
@@ -14,7 +14,6 @@ use wit_component::{ComponentEncoder, StringEncoding};
 use wit_parser::{Resolve, WorldId, WorldItem};
 
 mod flavorful;
-mod lists;
 mod many_arguments;
 mod options;
 mod other_dependencies;
